refactor: Group object-related utils under 📁`Util/Objects`
Redo of PR #3252. My git-fu is weak, and attempts to rebase and merge went fakakta
