+++ src/UserConfig.js
@@ -8,7 +8,7 @@ import EventEmitter from "./Util/AsyncEventEmitter.js";
 import EleventyCompatibility from "./Util/Compatibility.js";
 import EleventyBaseError from "./Errors/EleventyBaseError.js";
 import BenchmarkManager from "./Benchmark/BenchmarkManager.js";
-import { DeepCopy } from "./Util/Merge.js";
+import { DeepCopy } from "./Util/Objects/Merge.js";
 
 const debug = debugUtil("Eleventy:UserConfig");
